% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcesamientodDeDatos.R
\name{AgregaTasaPoblacional}
\alias{AgregaTasaPoblacional}
\title{Calcula tasa poblacional}
\usage{
AgregaTasaPoblacional(
  resumen_datos_estatal,
  poblacion_inegi_2015,
  columna_a_tasa = "casos_por_estado"
)
}
\arguments{
\item{resumen_datos_estatal}{Datos con los casos de violencia por estado}

\item{poblacion_inegi_2015}{Set de datos con la población total por estado}

\item{columna_a_tasa}{Cadena de tecto con el nombre de la olumna con los casos de violencia que se transforman a tasa}
}
\value{
Datos con dos columnas extras tasa por cada 100 mil habitantes y habitantes totales para ese estado.
}
\description{
Usando los datos de población por estado para el año 2015 se calcula la tasa
de casos por cada 100,000 habitantes y permitir una comparativa un tanto ponderada
}
\examples{

  AgregaTasaPoblacional(
    resumen_datos_estatal,
    poblacion_inegi_2015,
    columna_a_tasa = "casos_por_estado"
  )

}
