% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatosTransformados.R
\name{ComparaMesesConDatos}
\alias{ComparaMesesConDatos}
\title{Comparativa de Meses con Datos}
\usage{
ComparaMesesConDatos(datos_violencia, poblacion_inegi_2015, filtro.tipo = NULL)
}
\arguments{
\item{datos_violencia}{Conjunto de datos con el conteo de casos de violecina en méxico}

\item{poblacion_inegi_2015}{Población registrada para el año 2015}

\item{filtro.tipo}{Cadena de texto con el tipo de violencia a filtrar}
}
\value{
Genera un conjunto de datos que compara solo los meses con información
}
\description{
Genera un conjunto de datos con los meses que presentan datos en todos los
años existentes en el conjunto de datos. La ocurrencia de casos se representa con
la tasa por cada 100 mil habitantes. Ademas permite filtrar los casos de violencia
por el tipo.
}
\examples{
ComparaMesesConDatos(
  datos_violencia,
  poblacion_inegi_2015,
  filtro.tipo = "Violencia familiar"
)
}
