% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcesamientodDeDatos.R
\name{CreaPaletaPersonalizada}
\alias{CreaPaletaPersonalizada}
\title{Crae Paleta de colores personalizada}
\usage{
CreaPaletaPersonalizada(
  vector_usar,
  vector_colores,
  color_na = "#9C9C9C",
  numero_colores = 6
)
}
\arguments{
\item{vector_usar}{Vector de datos a los que se le genera un color asignado}

\item{vector_colores}{Vector de dos o mas colores usados para generar la escala de colores}

\item{color_na}{Color asignado a valores NA}

\item{numero_colores}{Variación de colores a usar para generar la escala final}
}
\value{
Función que genera una escala de colores
}
\description{
Esta función permite crear un vector con los colores entre 2 o 3 colores a
partir de un vector de datos numéricos para ser asignados al mismo.
}
\examples{
paleta.colores <- CreaPaletaPersonalizada(1:5, c("#2F394D", "#EEE1B3"))
}
