% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos_violencia}
\alias{datos_violencia}
\title{Datos de violencia en México}
\format{
Un tibble con 7 columnas y 225792 renglones:
\describe{
\item{Entidad}{Estado de la republica mexicana}
\item{Bien_juridico_afectado}{Los bienes involucrados en el eacto de violencia}
\item{Tipo}{Tipo de evento violento registrado}
\item{Subtipo}{Subtipo de violencia}
\item{Modalidad}{Modo con el que se llevo a cabo el acto de violencia registrada}
\item{ocurrencia}{Número de casos registrados para el estado de ese tipo, modalidad y fecha}
\item{fecha}{Fecha de registro del evento de violencia}
}
}
\usage{
datos_violencia
}
\description{
Este conjunto de datos presenta el conteo de casos para diferentes tipos de violencia,
con información de fecha, entidad, tipo, subtipo, modalidad y Bien Jurídico involucrado.
}
\keyword{datasets}
