% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatosTransformados.R
\name{DatosMesEstadoAgrupados}
\alias{DatosMesEstadoAgrupados}
\title{Datos de Meses por Estado Agrupados popr tipo}
\usage{
DatosMesEstadoAgrupados(
  datos_violencia,
  entidad = "Todas",
  resaltar.tipo = NULL
)
}
\arguments{
\item{datos_violencia}{Datos de violencia contra la mujer}

\item{entidad}{Entidad que se desea obtener, si se coloca "Todas" se obtienen los datos de toda la republica}

\item{resaltar.tipo}{Vector con el tipo de violencia que se mantendra, el resto se agrupa en un grupo.
Sí se coloca NULL se obtendran todos los tipos.}
}
\value{
Conjunto de datos con la Entidad seleccionada, El tipo de de violencia, año, mes,
casos registrados para el tipo de violencia, casos registrados totales para la entidad, mes y año.
Y porcentaje que este tipo representa.
}
\description{
Esta función genera un dataset con los datos de meses que estan completos en el todos los años.
Estos datos los muestra para un estado en específico. De seleccionar \strong{Todos} se mostraran los datos
de toda la republica. Ademas esta funcion permite seleccionar el tipo de violencia a la que se le va
a dar mayor relevancia. Es decir si la opcion \code{resaltar.tipo} esta a NULL se incluyen todos los tipos,
si se agrega un vector, estos tipos se mantendran y el resto de tipos de violencia se agruparan con la etiqueta
"Otros tipos de violencia".
}
\examples{
DatosMesEstadoAgrupados(
  datos_violencia,  entidad = "Todas", resaltar.tipo = c("Violencia familiar", "Abuso sexual")
)
}
