% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcesamientodDeDatos.R
\name{DefinirMesesSinDatos}
\alias{DefinirMesesSinDatos}
\title{Definir Meses sin datos registrados}
\usage{
DefinirMesesSinDatos(datos_violencia)
}
\arguments{
\item{datos_violencia}{Dataset de violencia procesado}
}
\value{
Vector con las fechas del tipo "2019-11-01"
}
\description{
Detecta los meses del dataset que no tienen registros de violencia, en
general puede deberse a años incompletos en información.
}
\examples{
DefinirMesesSinDatos(datos_violencia)
}
