% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EjecutaShinyApp.R
\name{ejecutaShyniApp}
\alias{ejecutaShyniApp}
\title{Ejecuta aplicación Shiny}
\usage{
ejecutaShyniApp()
}
\value{
Aplicación shiny en función
}
\description{
Ejecuta la aplicación shiny para la exploración de datos de
violencia contra la mujer.
}
\examples{
\dontrun{runShinyApp()}
}
