% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poblacion_inegi_2015}
\alias{poblacion_inegi_2015}
\title{Registro de población estatal 2015}
\format{
Un tibble con 2 columnas y 32 renglones:
\describe{
\item{Entidad}{Estado de la republica mexicana}
\item{Habitantes2015}{Número de habitantes registrados por el INEGI para el año 2015}
}
}
\usage{
poblacion_inegi_2015
}
\description{
Registro de la población por estado reportado por el INEGI para el año 2015
}
\keyword{datasets}
