% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poligonos_mx}
\alias{poligonos_mx}
\title{Poligonos de Estados de la Republica Mexicana}
\format{
Un objeto de clase sf con la relación de poligonos/entidad
\describe{
\item{id}{Identificador de registro}
\item{cvegeo}{Identificador de la entidad federativa}
\item{nombre}{Nombre de la entidad federativa}
\item{geometry}{Coordenadas geográficas que generan el poligono de la entidad}
}
}
\usage{
poligonos_mx
}
\description{
Este set de datos presenta los polígonos para los estados de la republica mexicana
}
\keyword{datasets}
