% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatosTransformados.R
\name{RankingTipoViolencia}
\alias{RankingTipoViolencia}
\title{Calculate ranking de tipo de violencia contra la mujer}
\usage{
RankingTipoViolencia(
  datos_violencia,
  entidad = NULL,
  numero_posiciones = NULL,
  resaltar_tipo = "Violencia familiar"
)
}
\arguments{
\item{datos_violencia}{Datos de casos de violencia contra la mujer}

\item{entidad}{Entidad para filtrar el dataset, en caso de no declararlo o ponerlo a nul se calcula a todo México}

\item{numero_posiciones}{Posiciones que se muestran dentro del ranking, por año y tipo de violencia.}

\item{resaltar_tipo}{Determina el tipo de violencia contra la mujer que se resaltará con un color en el plot}
}
\value{
Conjunto de datos con la columna, Tipo de caso, año, número de casos y psoición dentro del rank
}
\description{
Calcula la posición de cada tipo de violencia contra la mujer con base en la cuenta total de
casos por año por entidad, en caso de colocar a NULL la entidad o no declarar el parámetro se calcula
a total toda la republica mexicana.
}
\examples{
 RankingTipoViolencia(datos_violencia, entidad = "Querétaro", numero_posiciones = 5)
}
