% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatosTransformados.R
\name{TasaPromedioMensual}
\alias{TasaPromedioMensual}
\title{Tasa Promedio Mensual}
\usage{
TasaPromedioMensual(datos_violencia, poblacion_inegi_2015, filtro.tipo = NULL)
}
\arguments{
\item{datos_violencia}{Datos de violencia para la republica mexicana}

\item{poblacion_inegi_2015}{Población registrada para el año 2015}

\item{filtro.tipo}{Filtro del tipo de violencia ocurrida, en caso de ser NULL se
usan todos los tipos de violencia}
}
\value{
Regresa un conjunto de datos con la tasa promedio por cada 100mil habitantes al mes.
}
\description{
Calcula la tasa promedio mensual de los meses con información
dentro del conjunto de datos de violencia para México.
Permite hacer un filtrado por tipo de violencia.
}
\examples{
TasaPromedioMensual(datos_violencia, poblacion_inegi_2015, filtro.tipo = "Violencia familiar")
}
